backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.27.2
python-dotenv==1.0.1
itsdangerous==2.2.0
supabase==2.5.0
openai==1.59.7
pinecone-client==5.0.1
python-multipart==0.0.9
